+++ common/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "composite": true,
+    "outDir": "out",
+    "rootDir": ".",
+  },
+  "include": ["*.ts"]
+}
